requirements.txt
Django==1.7.4
gunicorn==19.3.0
djangorestframework==3.1.1
django-bower==5.0.3
django-imagekit==3.2.6
